feat: routes and view's folders were organized
