+++ src/app/app.routes.ts
@@ -1,10 +1,8 @@
 import { Routes } from '@angular/router';
-import {HomePetOwnerComponent} from "./views/home-pet-owner/home-pet-owner.component";
-import {CarePetOwnerComponent} from "./views/care-pet-owner/care-pet-owner.component";
+import {PET_OWNER_ROUTES} from "./views/pet-owner/pet-owner.routes";
+import {AUTH_ROUTES} from "./auth/auth.routes";
 
 export const routes: Routes = [
-  {path:"home/pet-owner", component:HomePetOwnerComponent},
-  {path:"care/pet-owner", component:CarePetOwnerComponent},
-  {path: '', loadChildren: () => import ('./auth/auth.routes').then(m => m.AUTH_ROUTES)}
-
+  ...AUTH_ROUTES,
+  ...PET_OWNER_ROUTES,
 ];
