+++ src/app/views/pet-owner/home-pet-owner/components/list-owner-pets/list-owner-pets.component.ts
@@ -1,11 +1,11 @@
 import {Component, ElementRef, Input, ViewChild} from '@angular/core';
 import {PetResponse} from "../../interfaces/PetResponse";
-import {PetCardComponent} from "../../../../shared/components/pet-card/pet-card.component";
 import {NgForOf} from "@angular/common";
 import {HomePetOwnerService} from "../../services/home-pet-owner.service";
 import {FormAddPetComponent} from "../form-add-pet/form-add-pet.component";
 import {DialogModule} from "primeng/dialog";
 import {TypeForm} from "../../interfaces/type-form.enum";
+import {PetCardComponent} from "../../../../../shared/components/pet-card/pet-card.component";
 
 @Component({
   selector: 'app-list-owner-pets',
