+++ src/app/views/pet-owner/home-pet-owner/home-pet-owner.component.html
@@ -1,0 +1,6 @@
+<div class="home-owner-content flex-grow p-5">
+      <h1>Home</h1>
+      <app-banner-owner></app-banner-owner>
+      <app-list-owner-pets></app-list-owner-pets>
+      <app-list-scpecialists-nearby></app-list-scpecialists-nearby>
+</div>
