+++ src/app/views/pet-owner/home-pet-owner/home-pet-owner.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import {PetCardComponent} from "../../shared/components/pet-card/pet-card.component";
 import {BannerOwnerComponent} from "./components/banner-owner/banner-owner.component";
 import {ListOwnerPetsComponent} from "./components/list-owner-pets/list-owner-pets.component";
 import {
@@ -8,9 +7,8 @@ import {
 import {HomePetOwnerService} from "./services/home-pet-owner.service";
 import {PetResponse} from "./interfaces/PetResponse";
 import {FormAddPetComponent} from "./components/form-add-pet/form-add-pet.component";
-import {SidebarComponent} from "../../shared/components/sidebar/sidebar.component";
 import {NgClass} from "@angular/common";
-import {NavBarComponent} from "../../shared/components/nav-bar/nav-bar.component";
+import {PetCardComponent} from "../../../shared/components/pet-card/pet-card.component";
 
 @Component({
   selector: 'app-home-pet-owner',
@@ -20,18 +18,11 @@ import {NavBarComponent} from "../../shared/components/nav-bar/nav-bar.component
     BannerOwnerComponent,
     ListOwnerPetsComponent,
     ListScpecialistsNearbyComponent,
-    SidebarComponent,
     NgClass,
-    NavBarComponent
   ],
   templateUrl: './home-pet-owner.component.html',
   styleUrl: './home-pet-owner.component.css'
 })
 export class HomePetOwnerComponent {
-  isSidebarCollapsed=false;
-
-  onSidebarToggle(isCollapsed:boolean){
-    this.isSidebarCollapsed = isCollapsed
-  }
 
 }
