+++ src/app/views/pet-owner/main-layout/main-layout.component.html
@@ -3,11 +3,7 @@
   <app-sidebar (sidebarToggle)="onSidebarToggle($event)" ></app-sidebar>
   <div [ngClass]="{'w-full': isSidebarCollapsed, 'w-4/5': !isSidebarCollapsed}" class="transition-width duration-300">
 
-    <div class="home-owner-content flex-grow p-5">
-      <h1>Home</h1>
-      <app-banner-owner></app-banner-owner>
-      <app-list-owner-pets></app-list-owner-pets>
-      <app-list-scpecialists-nearby></app-list-scpecialists-nearby>
-    </div>
+    <router-outlet></router-outlet>
+
   </div>
 </div>
